requirements.txt
selenium==4.23.1
requests==2.32.3
pytest==8.2.2
allure-pytest==2.13.5